refactor(chat, leave): update components and pages to use domain types (Phase 25c)
- Update chat-mention-popup to import MentionItem and MentionMode from domain layer
- Update dashboard-switcher to import LeaveRole from domain layer
- Update leaves management page to import LeaveRole from domain layer
